+++ src/components/book/book-fetching.tsx
@@ -5,35 +5,17 @@ import { useSession } from 'next-auth/react';
 import { useQuery } from '@tanstack/react-query';
 import { bookDetailOption } from './book-detail-option';
 import MobileBookDisplay from './mobile-book-display';
-import { useEffect, useState } from 'react';
+import { useIsMobile } from '@/hooks/useMediaQuery';
 
 export default function BookFetching({ slug }: { slug: string }) {
   const { data: session } = useSession();
-  const [isMobile, setIsMobile] = useState<boolean | null>(null);
+  const isMobile = useIsMobile();
 
   const {
     data: bookData,
     isLoading,
     error,
   } = useQuery(bookDetailOption(slug, session?.accessToken));
-
-  useEffect(() => {
-    const mediaQuery = window.matchMedia('(max-width: 767px)');
-
-    const handleChange = (e: MediaQueryListEvent | MediaQueryList) => {
-      setIsMobile(e.matches);
-    };
-
-    // 초기값 설정
-    handleChange(mediaQuery);
-
-    // 이벤트 리스너 등록
-    mediaQuery.addEventListener('change', handleChange);
-
-    return () => {
-      mediaQuery.removeEventListener('change', handleChange);
-    };
-  }, []);
 
   if (isLoading) return <div>책을 불러오는 중...</div>;
   if (!bookData) return <div>해당 책을 찾을 수 없습니다.</div>;
